Remove dictionary argument for OSPF module
This is pending a mailing list response. I want to be able to pass a dictionary to my OSPF module that represents a table of
neighbor information, but can't do it yet.

Let this commit represent the "old way", since the new way should be much more powerful. More on that later.