Worked around ROW issue
http://pastebin.com/GSwVzZyV

Basically if there's one area, then the value for a ROW tag in JSON doesn't show up as a list.
However, if there's multiple, then the value for a ROW tag in JSON DOES show up as a list.

So on the one hand, I refer to the tag directly, and on the other, I iterate through it. Ugh.

I have a few things commented out in nxapi_ospf that were a result of me fixing the errors that crop up because of this.
Wanted to make a commit before I took it all out.

Should go back and test all kinds of things, not just multiple areas.